Eliminate $TEST_DATA
This variable influenced too many files which don't belong together.
Replace its usage with more sensible defaults and better storage
behaviour:

 * Downloaded images: These are precious as a local developer, so put
   them into the XDG cache dir by default: ~/.cache/cockpit-images/.
   With that, there is no configuration needed, and one can clean git
   trees, have multiple checkouts etc. without problems. The path can
   still be changed with either a symlink or the
   cockpit.bots/images-data-dir git config.

   In CI the images are usually on a persistent external volume.
   ~/.cache/cockpit-images now *only* contains images, so it's easier to
   set up the volume without putting a lot of irrelevant crap into it
   (such as overlays, built rpms, lock files, etc.)

 * Ephemeral VM overlays: These should never be put into a container
   volume, and they should be on lightning fast storage (ideally, on
   tmpfs). By default, put them into tmp/run/ to keep the current
   behaviour (without `$TEST_DATA`), but allow bots to change the
   location with `$TEST_OVERLAY_DIR`.

 * GitHub cache: This is safe to share with other bots, but they are not
   particularly precious nor big. So we don't need to special-case them
   for CI. Keep them in ~/.cache/github/, there's no need to override
   that.

 * SSH master control sockets: These must *never ever* be shared with
   other bots. Put them into /tmp/.cockpit-test-resources/, where
   the image/port lock files already live.

Eliminate the usage of get_temp_dir(). It's too unspecific and can't fit
all the diverging use cases above. Still keep a backwards compatible
shim until cockpit drops its usage in all of its branches (see
https://github.com/cockpit-project/cockpit/pull/14190).

Closes #943
