+++ machine/machine_core/ssh_connection.py
@@ -23,10 +23,10 @@ import subprocess
 import select
 import errno
 import sys
+import tempfile
 
 from . import exceptions
 from . import timeout as timeoutlib
-from .directories import get_temp_dir
 
 
 class SSHConnection(object):
@@ -135,7 +135,8 @@ class SSHConnection(object):
     def _start_ssh_master(self):
         self._kill_ssh_master()
 
-        control = os.path.join(get_temp_dir(), "ssh-%h-%p-%r-" + str(os.getpid()))
+        control = os.path.join(tempfile.gettempdir(), ".cockpit-test-resources", "ssh-%h-%p-%r-" + str(os.getpid()))
+        os.makedirs(os.path.dirname(control), exist_ok=True)
 
         cmd = [
             "ssh",
